+++ include/Isen/NameList.h
@@ -50,17 +50,17 @@ struct NameList
     /// Horizontal resolution
     int nx = 100;
     /// Horizontal spacing
-    Float dx = xl / nx;
+    double dx = xl / nx;
     /// Domain depth [K]
-    Float thl = 60.0;
+    double thl = 60.0;
     /// Vertical Resolution
     int nz = 60;
     /// Integration time [s]
-    Float time = 6 * 60 * 60;
+    double time = 6 * 60 * 60;
     /// Time step [s]
-    Float dt = 10;
+    double dt = 10;
     /// Horizontal diffusion coefficient
-    Float diff = 0.02;
+    double diff = 0.02;
 
     //-------------------------------------------------
     // Topography
@@ -78,11 +78,11 @@ struct NameList
     //-------------------------------------------------
 
     /// Initial velocity [m/s]
-    Float u00 = 15;
+    double u00 = 15;
     /// Brunt - Vaisalla frequency [1/s]
-    Float bv00 = 0.01;
+    double bv00 = 0.01;
     /// Potential temperature at surface
-    Float th00 = 300;
+    double th00 = 300;
     /// Wind shear simulation
     bool ishear = false;
     /// Bottom level of wind shear layer (bottom level of wind layer is 0)
@@ -90,7 +90,7 @@ struct NameList
     /// Top level of wind shear layer (top level of wind layer is nz - 1)
     int k_sht = 8;
     /// Initial velocity below shear layer [m/s]
-    Float u00_sh = 10;
+    double u00_sh = 10;
 
     //-------------------------------------------------
     // Boundaries
@@ -99,7 +99,7 @@ struct NameList
     /// Number of grid points in absorber
     int nab = 0;
     /// Maximum value of absorber
-    Float diffabs = 1.0;
+    double diffabs = 1.0;
     /// Lateral boundaries (false = periodic)
     bool irelax = false;
     /// Number of boundary points on each side
@@ -136,11 +136,11 @@ struct NameList
     //-------------------------------------------------
 
     /// Multiplication factor for terminal fall velocity
-    Float vt_mult = 1.0;
+    double vt_mult = 1.0;
     /// Critical cloud water mixing ratio for the onset of autoconversion [kg/kg]
-    Float autoconv_th = 0.0001;
+    double autoconv_th = 0.0001;
     /// Multiplication factor for autoconversion
-    Float autoconv_mult = 1.0;
+    double autoconv_mult = 1.0;
     /// Switch to turn on / off sedimentation
     bool sediment_on = 1;
 
@@ -149,7 +149,7 @@ struct NameList
     //-------------------------------------------------
 
     /// Spacing between vertical layers[K]
-    Float dth = thl / nz;
+    double dth = thl / nz;
     /// Number of iterations
     int nts = static_cast<int>(std::round(time / dt));
     /// Number of output steps
@@ -178,25 +178,25 @@ struct NameList
     //-------------------------------------------------
 
     /// Gravity
-    const Float g = 9.81;
+    const double g = 9.81;
     /// Specific heat of air at constant pressure
-    const Float cp = 1004.0;
+    const double cp = 1004.0;
     /// Gas constant of air [J/kgK]
-    const Float r = 287.0;
+    const double r = 287.0;
     /// Gas constant of vapor [J/kgK]
-    const Float r_v = 461.0;
+    const double r_v = 461.0;
     /// Short cut for R / Cp
-    const Float rdcp = r / cp;
+    const double rdcp = r / cp;
     /// short cut for Cp / R
-    const Float cpdr = cp / r;
+    const double cpdr = cp / r;
     /// Reference pressure in SI units(Pa, not hPa!)
-    const Float pref = 100 * 1000;
+    const double pref = 100 * 1000;
     /// Surface height
-    const Float z00 = 0.0;
+    const double z00 = 0.0;
     /// Upstream surface pressure (= reference pressure)
-    const Float prs00 = pref;
+    const double prs00 = pref;
     ///
-    const Float exn00 = cp * std::pow(prs00 / pref, rdcp);
+    const double exn00 = cp * std::pow(prs00 / pref, rdcp);
 
     //-------------------------------------------------
 
@@ -205,7 +205,7 @@ struct NameList
 
     /// Set a variable by name [potentially slow]
     void setByName(const std::string& name, const int& value);
-    void setByName(const std::string& name, const Float& value);
+    void setByName(const std::string& name, const double& value);
     void setByName(const std::string& name, const bool& value);
     void setByName(const std::string& name, const std::string& value);
